erase unnecessary frees and exits
If it is the left child, it will call exec_cmd with cmd being an execcmd or a backcmd, and that will end in one of the cases EXEC, REDIR or BACK. If execvp fails the error will be handled in those places.

If it is the right child we have two options. Either it is the last in the pipeline, so the same for the left child applies. Or, it is a new coordinator process, who will free the parsed command and the alternative stack after both children have died.

Co-Authored-By: Camilo Sassone Irrazabal <40524371+Vityyy@users.noreply.github.com>
Co-Authored-By: Valentín Bersi <valentinbersi@gmail.com>
Co-Authored-By: Tomás Hevia <118986985+tomashevia27@users.noreply.github.com>